fixed safeties and isUsingPID issue
I fixed the simple issue of PID not running when manual control was disabled. I also fixed the issue where manual control would not go down if it was at the top limit and would not go up if it was at the bottom limit. Essentially, if any limit was hit the sprocket would freeze up even if moving down was safe or moving up was safe. I commented out the methods that move the sprocket using the DPAD, and I will delete them once I make sure the joystick commands are working. Long message!!

toodles!