Forgot to stage files last time!
Sorry, I woke up in the middle of the night and decided to check my code and saw that in my last commit I added nonsense. This is because I forgot to stage my files! I lost my progress but I think it might be fixed now. I don't know it's the middle of the night.